Move common files to be used by both extensions
