+++ firefox-ios/Client/Utils/ExtensionsUtils/SendToDevice.swift
