+++ firefox-ios/Client/Utils/ExtensionsUtils/ShareViewController.swift
