+++ firefox-ios/Client/Utils/ExtensionsUtils/UXConstants.swift
